Fix mangled strings in toplevel parse
It looks like the compiler and PARSE were using the same transient region for storing source code and strings, resulting in mangled strings when you have a newline between PARSE and e.g. TYPE, which is confusing and obviously undesired. Automatically using SAVE-MEM copies the parsed string to a more stable part of the host memory so the string is persisted correctly.